+++ src/protocols/index.ts
@@ -1,0 +1,2 @@
+export * from './AboutChannelProtocol/AboutChannelProtocol.js';
+export * from './PlaylistArrayProtocol/PlaylistArrayProtocol.js';