+++ CHANGELOG.md
@@ -7,6 +7,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
-## 1.0.1 - 2024-01-07
 ### Added
+
+- Added protocol named **ChannelDetails** to fetch channel details such as name, title, description, videos count etc.
+- Added protocol named **PlaylistArray** to fetch list of playlists available on a channel.
+
+## 1.0.1 - 2024-01-07
+
+### Added
+
 - Changelog